+++ src/index.html
@@ -128,12 +128,11 @@
   </a>
 </template>
 
-
+<script src="js/loader.js" charset="utf-8"></script>
+<script src="js/get-picture-element.js" charset="utf-8"></script>
 <script src="js/resizer.js" charset="utf-8"></script>
 <script src="js/upload.js" charset="utf-8"></script>
-<!--<script src="js/check.js" charset="utf-8"></script>
-<script src="js/upload-stat.min.js" charset="utf-8"></script>-->
-<script src="pictures.js" charset="utf-8"></script>
+<script src="js/pictures.js" charset="utf-8"></script>
 
 </body>
 </html>
