+++ src/js/get-picture-element.js
@@ -1,14 +1,8 @@
 'use strict';
-
-var urlToLoad = 'http://localhost:1507/api/pictures';
-var filterBlock = document.querySelector('.filters');
-var container = document.querySelector('.pictures');
 var template = document.querySelector('#picture-template');
 var templateContainer = 'content' in template ? template.content : template;
 var pictureElementNode = templateContainer.querySelector('.picture');
 var pictureSide = '182';
-
-filterBlock.classList.add('hidden');
 
 var getPictureElement = function(picture) {
   var pictureElement = pictureElementNode.cloneNode(true);
@@ -29,27 +23,3 @@ var getPictureElement = function(picture) {
   backgroundImage.src = picture.url;
   return pictureElement;
 };
-
-var loadPictures = function(url, callback, callbackName) {
-  if (!callbackName) {
-    callbackName = 'mcb' + Date.now();
-  }
-  window[callbackName] = function(dataFromServer) {
-    callback(dataFromServer);
-    script.remove();
-  };
-  var script = document.createElement('script');
-  script.src = url + '?callback=' + callbackName;
-  document.body.appendChild(script);
-};
-
-var renderPictures = function(picArray) {
-  picArray.forEach(function(picture) {
-    container.appendChild(getPictureElement(picture));
-  });
-  filterBlock.classList.remove('hidden');
-};
-
-loadPictures(urlToLoad, function(dataFromServer) {
-  renderPictures(dataFromServer);
-});
