+++ src/js/render.js
@@ -1,0 +1,8 @@
+'use strict';
+
+var renderPictures = function(picArray) {
+  picArray.forEach(function(picture) {
+    container.appendChild(getPictureElement(picture));
+  });
+  filterBlock.classList.remove('hidden');
+};
